rename package to migrator
