+++ src/main/java/org/betonquest/betonquest/config/patcher/migration/migrator/from1to2/EventScheduling.java
@@ -1,7 +1,8 @@
-package org.betonquest.betonquest.config.patcher.migration.migrators.from1to2;
+package org.betonquest.betonquest.config.patcher.migration.migrator.from1to2;
 
 import org.betonquest.betonquest.config.patcher.migration.FileConfigurationProvider;
 import org.betonquest.betonquest.config.patcher.migration.Migration;
+import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.YamlConfiguration;
 
 import java.io.File;
@@ -9,14 +10,9 @@ import java.io.IOException;
 import java.util.Map;
 
 /**
- * Handles the PackageSection migration.
+ * Handles the EventScheduling migration.
  */
-public class PackageSection implements Migration {
-
-    /**
-     * The enabled string.
-     */
-    public static final String ENABLED = "enabled";
+public class EventScheduling implements Migration {
 
     /**
      * The config producer.
@@ -24,11 +20,11 @@ public class PackageSection implements Migration {
     private final FileConfigurationProvider producer;
 
     /**
-     * Creates a new PackageSection migrator.
+     * Creates a new EventScheduling migrator.
      *
      * @param provider The config provider
      */
-    public PackageSection(final FileConfigurationProvider provider) {
+    public EventScheduling(final FileConfigurationProvider provider) {
         this.producer = provider;
     }
 
@@ -38,10 +34,14 @@ public class PackageSection implements Migration {
         for (final Map.Entry<File, YamlConfiguration> entry : configs.entrySet()) {
             final File file = entry.getKey();
             final YamlConfiguration config = entry.getValue();
-            if (config.contains(ENABLED, true)) {
-                final boolean section = config.getBoolean(ENABLED);
-                config.set("package.enabled", section);
-                config.set(ENABLED, null);
+            final ConfigurationSection staticSection = config.getConfigurationSection("static");
+            if (staticSection != null) {
+                staticSection.getValues(false).forEach((key, value) -> {
+                    config.set("schedules." + key + ".type", "realtime-daily");
+                    config.set("schedules." + key + ".time", key);
+                    config.set("schedules." + key + ".events", value);
+                });
+                config.set("static", null);
                 config.save(file);
             }
         }
